OCT-176: use constraint in Catalog domain class
