+++ components/catalogs/back/src/Infrastructure/Validation/CatalogProductMapping.php
@@ -12,6 +12,7 @@ use Symfony\Component\Validator\Constraint;
  *
  * @psalm-suppress PropertyNotSetInConstructor
  */
-final class CatalogUpdateProductValueFiltersPayload extends Constraint
+#[\Attribute]
+final class CatalogProductMapping extends Constraint
 {
 }
