+++ components/catalogs/back/src/Infrastructure/Validation/CatalogProductSelectionCriteria.php
@@ -12,6 +12,7 @@ use Symfony\Component\Validator\Constraint;
  *
  * @psalm-suppress PropertyNotSetInConstructor
  */
-final class CatalogUpdatePayload extends Constraint
+#[\Attribute]
+final class CatalogProductSelectionCriteria extends Constraint
 {
 }
