+++ components/catalogs/back/src/Infrastructure/Validation/CatalogProductSelectionCriteriaValidator.php
@@ -30,7 +30,7 @@ use Symfony\Component\Validator\Exception\UnexpectedTypeException;
  *
  * @psalm-suppress PropertyNotSetInConstructor
  */
-final class CatalogUpdateProductSelectionCriteriaPayloadValidator extends ConstraintValidator
+final class CatalogProductSelectionCriteriaValidator extends ConstraintValidator
 {
     public function __construct(
         private readonly FindOneAttributeByCodeQueryInterface $findOneAttributeByCodeQuery,
@@ -40,8 +40,8 @@ final class CatalogUpdateProductSelectionCriteriaPayloadValidator extends Constr
 
     public function validate($value, Constraint $constraint): void
     {
-        if (!$constraint instanceof CatalogUpdateProductSelectionCriteriaPayload) {
-            throw new UnexpectedTypeException($constraint, CatalogUpdateProductSelectionCriteriaPayload::class);
+        if (!$constraint instanceof CatalogProductSelectionCriteria) {
+            throw new UnexpectedTypeException($constraint, CatalogProductSelectionCriteria::class);
         }
 
         $this->context
@@ -53,7 +53,7 @@ final class CatalogUpdateProductSelectionCriteriaPayloadValidator extends Constr
     /**
      * @return array<array-key, Constraint>
      */
-    private function getConstraints(CatalogUpdateProductSelectionCriteriaPayload $constraint): array
+    private function getConstraints(CatalogProductSelectionCriteria $constraint): array
     {
         $maxCriteriaPerCatalog = $this->maxCriteriaPerCatalog;
         return [
