--- components/catalogs/back/src/Infrastructure/Validation/CatalogUpdateProductSelectionCriteriaPayload.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Akeneo\Catalogs\Infrastructure\Validation;
-
-use Symfony\Component\Validator\Constraint;
-
-/**
- * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @psalm-suppress PropertyNotSetInConstructor
- */
-final class CatalogUpdateProductSelectionCriteriaPayload extends Constraint
-{
-}
